% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1_delete.R
\name{d1_delete}
\alias{d1_archive}
\alias{d1_delete}
\title{d1_delete}
\usage{
d1_delete(id, cert = "/tmp/x509up_u1000",
  node = "https://knb.ecoinformatics.org/knb/d1/mn/v1")
}
\arguments{
\item{id}{the permanent identifier of the object}

\item{cert}{path to the x509 certificate from https://cilogon.org/?skin=DataONE}

\item{node}{the authoritativeNode where the data has been archived}
}
\description{
Archive a data object so that it is removed from indexing
}
\details{
To generate a valid certificate, visit https://cilogon.org/?skin=DataONE and download a certificate.  (If asked to execute or download `shibCILaunchGSCA.jnlp`, choose execute, which will launch the download.  Otherwise, download this file and run `javaws shibCILaunchGSCA.jnlp` on the command line to launch.  Requires a working Java webscript application.
}

