% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1_get.R
\name{d1_get}
\alias{d1_download}
\alias{d1_get}
\title{d1_get}
\usage{
d1_get(id, action = c("object", "meta"),
  node = c("https://cn.dataone.org/cn/v1",
  "https://knb.ecoinformatics.org/knb/d1/mn/v1/"))
}
\arguments{
\item{id}{the identifier of the object of interest}

\item{action}{return the object itself (`object`) or the system metadata (`meta`)? Defaults to `object`}

\item{node}{the node url of the DataONE node.  Be default, uses the central dataone.org node.  Usually no need to adjust this.}
}
\value{
httr "response" object.  Use httr function \code{\link{content}} to parse
}
\description{
get data or metadata given an identifier
}
\details{
action should be object to download an object itself, including a metadata object (such as an EML xml file).
 meta allows access to system metadata.
}
\examples{
\dontrun{
d1_get("doi:10.5063/AA/wolkovich.24.1")
}
}

