% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1_info.R
\name{d1_info}
\alias{d1_info}
\title{d1_info}
\usage{
d1_info(node = c("https://cn.dataone.org/cn/v1",
  "https://knb.ecoinformatics.org/knb/d1/mn/v1/"))
}
\arguments{
\item{node}{the URL of the DataONE node of interest.}
}
\value{
XML-noded description of the node in an httr response object.
}
\description{
Get information about the services available at a node.
}

