% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1_search.R
\name{d1_search}
\alias{d1_search}
\title{d1_search}
\usage{
d1_search(query, search_field, filter = NULL,
  return_fields = c("identifier"), fuzzy = TRUE, rows = 100,
  format = c("xml", "json", "csv"), node = "https://cn.dataone.org/cn/v1",
  engine = "solr")
}
\arguments{
\item{query}{the search query (character string)}

\item{search_field}{named string providing optional filters for the search,
e.g. c(author = "Wolkovich", identifier = "doi"),}

\item{filter}{filters to apply (depends on engine, e.g. solr filters)}

\item{return_fields}{fields that should be returned by the search, see details.}

\item{fuzzy}{logical indicating whether fuzzy or only exact matches should be returned.}

\item{rows}{maximum number of search results}

\item{format}{perferred fromat for search results}

\item{node}{which dataone node to search, defaults to master node.}

\item{engine}{search engine to use for queries, defaults to solr}
}
\description{
Search the dataone repository by fields
}
\details{
The current possible query terms for both search and return fields are given here: https://mule1.dataone.org/ArchitectureDocs-current/design/SearchMetadata.html#attribute-descriptions-and-notes
}
\examples{
\dontrun{
 d1_search("Wolkovich", "author")
}
}

