% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1_update.R
\name{d1_update}
\alias{d1_update}
\title{d1_update}
\usage{
d1_update(pid, object, uid, new_id = getid("extract", object),
  cert = "/tmp/x509up_u1000",
  node = "https://knb.ecoinformatics.org/knb/d1/mn/v1",
  sysmeta = write_sysmeta(object, uid = uid, id = new_id))
}
\arguments{
\item{pid}{package id to be updated}

\item{object}{new data file to be uploaded}

\item{uid}{the user id of the data maintainer}

\item{new_id}{new package id to be assigned}

\item{cert}{path to the x509 certificate from https://cilogon.org/?skin=DataONE}

\item{node}{The URL to the DataONE node we intend to update. Defaults to the KNB}

\item{sysmeta}{the required system metadata for the package, geranted by default.}
}
\value{
httr::response object indicating the success or failure of the call
}
\description{
Update an existing dataone entry
}
\examples{
\dontrun{
f <- system.file("doc", "reml_example.xml", package="reml")
d1_upload(f, "boettiger", id=uuid::UUIDgenerate(), node = knb_test)
change f

d1_update(
}
}

