% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d1_upload.R
\name{d1_upload}
\alias{d1_upload}
\title{d1_upload}
\usage{
d1_upload(object, uid, id = getid("extract", object),
  cert = "/tmp/x509up_u1000",
  node = "https://knb.ecoinformatics.org/knb/d1/mn/v1",
  sysmeta = write_sysmeta(object, uid = uid, id = id))
}
\arguments{
\item{object}{new data file to be uploaded}

\item{uid}{the user id of the data maintainer}

\item{id}{what identifier should be used for the object; default will try and guess from object metadata (e.g. EML metadata).}

\item{cert}{path to the x509 certificate from https://cilogon.org/?skin=DataONE}

\item{node}{The URL to the DataONE node we intend to update. Defaults to the KNB}

\item{sysmeta}{the required system metadata for the package, geranted by default.}
}
\value{
httr::response object indicating the success or failure of the call
}
\description{
upload an object to dataone
}
\examples{
\dontrun{
f <- system.file("doc", "reml_example.xml", package="reml")
d1_upload(f, "boettiger", id=uuid::UUIDgenerate(), node = knb_test)
}
}

