% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_sysmeta.R
\name{getid}
\alias{getid}
\title{getid}
\usage{
getid(id, file = NULL)
}
\arguments{
\item{id}{if "extract" or "create", will attempt to do so.}

\item{file}{file need only be provided for extract}
}
\description{
helper function to the get the id of a dataone object
}

