% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_sysmeta.R
\name{write_sysmeta}
\alias{write_sysmeta}
\title{write_sysmeta}
\usage{
write_sysmeta(file, uid, id = c("create", "extract"), formatId = "guess",
  affiliation = "unaffiliated", metafile = "sysmeta.xml")
}
\arguments{
\item{file}{the path/name of the file to generate metadata for.}

\item{uid}{your user identifier as registered at cilogon.org}

\item{id}{how to obtain the identifier.  If "create", will attempt to make an
identifier with uuid.  "extract" will attempt to extract EML packageID
identifiers.  Otherwise, just provide the identifier here.}

\item{formatId}{if "guess", will base a guess on the file extension, otherwise specify format here}

\item{affiliation}{affiliation as registered at http://cilogon.org certificate}

\item{metafile}{the name of the output minimal system metadata file (NULL returns output to console)}
}
\value{
an XML object containing the dataone minimal system metadata, see schema definition for details (in xsd directory)
}
\description{
Generate minimal system metadata XML file
}

